pyproject.toml
[tool.poetry]
name = "gmail_digest"
version = "1.3.1"
description = "Simple tool to take a CSV and a template and create drafts in your Gmail. Can be used as a command line tool, or as a library."
authors = ["Michael Bianco <mike@mikebian.co>"]
license = "MIT"
readme = "readme.md"
homepage = "https://github.com/iloveitaly/gmail-draft-creator"
keywords = ["gmail", "email", "gmail-api"]

[project.urls]
'Changelog' = 'https://github.com/iloveitaly/gmail-draft-creator/blob/master/CHANGELOG.md'

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.137.0"
python-decouple = "^3.8"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
click = "^8.1.7"
markdown = "^3.6"
funcy-pipe = "^0.11.0"
openai = "^1.37.0"
structlog = "^24.4.0"
apscheduler = "^3.10.4"

[tool.poetry.scripts]
gmail-digest = "gmail_digest:main"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
exclude = '''
/(
  .venv
  | node_modules
  | migrations
)/
'''

# when changing this config, you need to restart your vscode isort server
[tool.isort]
profile = "black"
float_to_top = true
multi_line_output = 3
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
